+++ fidget/src/rhai/mod.rs
@@ -53,7 +53,9 @@ use std::sync::{Arc, Mutex};
 use crate::{context::Tree, Error};
 use rhai::{CustomType, EvalAltResult, NativeCallContext, TypeBuilder};
 
-pub mod shapes;
+mod shapes;
+mod tree;
+mod vec;
 
 /// Engine for evaluating a Rhai script with Fidget-specific bindings
 pub struct Engine {
@@ -77,62 +79,15 @@ impl Engine {
     /// which is effectively our standard library.
     pub fn new() -> Self {
         let mut engine = rhai::Engine::new();
-        engine
-            .register_type::<Tree>()
-            .register_fn("remap_xyz", remap_xyz);
+
+        tree::register(&mut engine);
+        vec::register(&mut engine);
+        shapes::register(&mut engine);
 
         engine.build_type::<Axes>();
         engine.register_fn("axes", axes);
         engine.register_fn("draw", draw);
         engine.register_fn("draw_rgb", draw_rgb);
-
-        macro_rules! register_binary_fns {
-            ($op:literal, $name:ident, $engine:ident) => {
-                $engine.register_fn($op, $name::node_dyn);
-                $engine.register_fn($op, $name::dyn_node);
-            };
-        }
-        macro_rules! register_unary_fns {
-            ($op:literal, $name:ident, $engine:ident) => {
-                $engine.register_fn($op, $name::node);
-            };
-        }
-
-        register_binary_fns!("+", add, engine);
-        register_binary_fns!("-", sub, engine);
-        register_binary_fns!("*", mul, engine);
-        register_binary_fns!("/", div, engine);
-        register_binary_fns!("%", modulo, engine);
-        register_binary_fns!("min", min, engine);
-        register_binary_fns!("max", max, engine);
-        register_binary_fns!("compare", compare, engine);
-        register_binary_fns!("and", and, engine);
-        register_binary_fns!("or", or, engine);
-        register_binary_fns!("atan2", atan2, engine);
-        register_unary_fns!("abs", abs, engine);
-        register_unary_fns!("sqrt", sqrt, engine);
-        register_unary_fns!("square", square, engine);
-        register_unary_fns!("sin", sin, engine);
-        register_unary_fns!("cos", cos, engine);
-        register_unary_fns!("tan", tan, engine);
-        register_unary_fns!("asin", asin, engine);
-        register_unary_fns!("acos", acos, engine);
-        register_unary_fns!("atan", atan, engine);
-        register_unary_fns!("exp", exp, engine);
-        register_unary_fns!("ln", ln, engine);
-        register_unary_fns!("not", not, engine);
-        register_unary_fns!("ceil", ceil, engine);
-        register_unary_fns!("floor", floor, engine);
-        register_unary_fns!("round", round, engine);
-        register_unary_fns!("-", neg, engine);
-
-        // Ban comparison operators
-        for op in ["==", "!=", "<", ">", "<=", ">="] {
-            engine.register_fn(op, bad_cmp_node_dyn);
-            engine.register_fn(op, bad_cmp_dyn_node);
-        }
-
-        engine.set_fast_operators(false);
 
         let context = Arc::new(Mutex::new(ScriptContext::new()));
         engine.set_default_tag(rhai::Dynamic::from(context.clone()));
@@ -175,19 +130,14 @@ impl Engine {
 
     /// Evaluates a single expression, in terms of `x`, `y`, and `z`
     pub fn eval(&mut self, script: &str) -> Result<Tree, Error> {
+        self.context.lock().unwrap().clear();
+
+        let mut scope = rhai::Scope::new();
+        scope.push("x", Tree::x());
+        scope.push("y", Tree::y());
+        scope.push("z", Tree::z());
+
         let ast = self.engine.compile(script)?;
-        let mut scope = {
-            let mut ctx = self.context.lock().unwrap();
-            ctx.clear();
-
-            // Create initialized scope with x/y/z
-            let mut scope = rhai::Scope::new();
-            scope.push("x", Tree::x());
-            scope.push("y", Tree::y());
-            scope.push("z", Tree::z());
-            scope
-        };
-
         let out = self.engine.eval_ast_with_scope::<Tree>(&mut scope, &ast)?;
 
         Ok(out)
@@ -250,10 +200,6 @@ fn axes(_ctx: NativeCallContext) -> Axes {
     Axes { x, y, z }
 }
 
-fn remap_xyz(shape: Tree, x: Tree, y: Tree, z: Tree) -> Tree {
-    shape.remap_xyz(x, y, z)
-}
-
 fn draw(
     ctx: NativeCallContext,
     tree: rhai::Dynamic,
@@ -291,95 +237,6 @@ fn draw_rgb(
     });
     Ok(())
 }
-
-macro_rules! define_binary_fns {
-    ($name:ident $(, $op:ident)?) => {
-        mod $name {
-            use super::*;
-            use NativeCallContext;
-            $(
-            use std::ops::$op;
-            )?
-            pub fn node_dyn(
-                ctx: NativeCallContext,
-                a: Tree,
-                b: rhai::Dynamic,
-            ) -> Result<Tree, Box<rhai::EvalAltResult>> {
-                let b = Tree::from_dynamic(&ctx, b)?;
-                Ok(a.$name(b))
-            }
-            pub fn dyn_node(
-                ctx: NativeCallContext,
-                a: rhai::Dynamic,
-                b: Tree,
-            ) -> Result<Tree, Box<rhai::EvalAltResult>> {
-                let a = Tree::from_dynamic(&ctx, a)?;
-                Ok(a.$name(b))
-            }
-        }
-    };
-}
-
-macro_rules! define_unary_fns {
-    ($name:ident) => {
-        mod $name {
-            use super::*;
-            pub fn node(
-                ctx: NativeCallContext,
-                a: rhai::Dynamic,
-            ) -> Result<Tree, Box<EvalAltResult>> {
-                let a = Tree::from_dynamic(&ctx, a)?;
-                Ok(a.$name())
-            }
-        }
-    };
-}
-
-fn bad_cmp_node_dyn(
-    _ctx: NativeCallContext,
-    _a: Tree,
-    _b: rhai::Dynamic,
-) -> Result<Tree, Box<rhai::EvalAltResult>> {
-    let e = "cannot compare Tree types during function tracing";
-    Err(e.into())
-}
-
-fn bad_cmp_dyn_node(
-    _ctx: NativeCallContext,
-    _a: rhai::Dynamic,
-    _b: Tree,
-) -> Result<Tree, Box<rhai::EvalAltResult>> {
-    let e = "cannot compare Tree types during function tracing";
-    Err(e.into())
-}
-
-define_binary_fns!(add, Add);
-define_binary_fns!(sub, Sub);
-define_binary_fns!(mul, Mul);
-define_binary_fns!(div, Div);
-define_binary_fns!(min);
-define_binary_fns!(max);
-define_binary_fns!(compare);
-define_binary_fns!(modulo);
-define_binary_fns!(and);
-define_binary_fns!(or);
-define_binary_fns!(atan2);
-define_unary_fns!(sqrt);
-define_unary_fns!(square);
-define_unary_fns!(neg);
-define_unary_fns!(sin);
-define_unary_fns!(cos);
-define_unary_fns!(tan);
-define_unary_fns!(asin);
-define_unary_fns!(acos);
-define_unary_fns!(atan);
-define_unary_fns!(exp);
-define_unary_fns!(ln);
-define_unary_fns!(not);
-define_unary_fns!(abs);
-define_unary_fns!(floor);
-define_unary_fns!(ceil);
-define_unary_fns!(round);
 
 ////////////////////////////////////////////////////////////////////////////////
